# database/base.py - Updated `__all__` attribute with type hinting.
# database/schema.sql
- Changed default values for `settings` table columns from NULL to `0`.
-- Added `personal_intros_channel_id`, `roles_channel_id` and `infraction_log_channel_id`.
- Added `role_embeds` table to handle Reaction based roles.

# database/settings.py
- Updated `__all__` attribute for `Role_Embed_Info` dataclass.
- Added `Role_Embed_Info(Base)` dataclass to handle database data.
-- Added `add_role_embeds()`.
-- Added `remove_role_embeds()`.
-- Added `get_all_role_embeds()`.
-- Added `get_role_embed()`.

# database/user.py
- Updated `__all__` attribute for `Image` dataclass.
- Improved login inside `remove_infraction()`.

# pnwbot/main.py
- Added `_bot_name` attribute.
- Set `self.owner_id = None`.
- Updated type hinting for `self._guild_settings` removing `None`.
- Added `on_member_update()` logic to handle verification roles and prompts for new members.
- Added prefix commands.
- Updated logic inside `sync()` command.

# pnwbot/infractions.py
- Init of file.
- Added logic for adding,removing and getting infractions for users.

# pnwbot/autorole.py
- Added `RoleButton(Button)` class for our View Buttons.
- Added `ReactionRoleView(View)` to handle our reaction role buttons.
- Added Role Group attributes broken into groups based on type.
- Added `autocomplete_role_embeds()` to fetch and handle Database embed information.
- Added `autocomplete_embed_buttons()` to fetch buttons related to an existing embed.
- Modified `role_embed()` command to support 5 roles and create buttons respectively.
- Added `add_button()` to allow adding of role buttons to existing role embeds.
- Added `remove_button()` to remove role buttons from existing role embeds.
- Updated `on_reaction_role()` to support more roles and grouping of roles to prevent duplicate entries.

+++ database/user.py
@@ -10,7 +10,7 @@ import util.asqlite as asqlite
 
 from .base import DB_FILE_PATH, Base
 
-__all__: tuple[str, str, str] = ("User", "Leave", "Infraction")
+__all__: tuple[str, ...] = ("User", "Leave", "Infraction", "Image",)
 
 
 @dataclass
@@ -184,10 +184,20 @@ class User(Base):
         return set([Infraction(**row) for row in res])
 
     @exists
-    async def remove_infraction(self, infraction: Infraction) -> set[Infraction]:
+    async def remove_infraction(self, infraction: Infraction | None = None, id: int | None = None) -> set[Infraction]:
+        if infraction is None and id is None:
+            raise ValueError("Either infraction or id must be provided")
+
+        if infraction is not None:
+            id = infraction.id
+
         await self._execute(SQL=f"""DELETE FROM infractions WHERE id=?""",
-                            parameters=(infraction.id,))
-        self.user_infractions.remove(infraction)
+                            parameters=(id,))
+
+        for infraction in self.user_infractions:
+            if infraction.id == id:
+                self.user_infractions.remove(infraction)
+
         return self.user_infractions
 
     @exists
